refactor(tests): split tests/test_asr.py into per-backend files
@geoHeil requested that the file be split rather than added to the max-lines ignore list. The previous tests/test_asr.py grew to 1094 lines after the regression tests in dea19a0 pushed it past the 1000-line limit introduced by the new max-lines pre-commit check.

Split into two files matching the upstream naming convention used for other ML test suites: tests/test_asr_mlx_whisper.py (364 lines) holds TestMlxWhisperIntegration, and tests/test_asr_whisper_s2t.py (740 lines) holds the four WhisperS2T test classes. Each file declares its own module-level pytestmark = pytest.mark.ml_asr so the marker-based CI selection still picks them up.

Also remove tests/test_asr.py from .github/max-lines-ignore now that the proper fix is in place. The exemption added in ad7f85a is no longer needed.
